Changed relative to absolute imports, commented out some imports that are not used in preparation for future removal, moved tests to new location
