+++ xpsi/Background.py
@@ -1,11 +1,11 @@
 
 __all__ = ["Background", "BackgroundError"]
 
-from .global_imports import *
-from . import global_imports
+from xpsi.global_imports import *
+#from . import global_imports
 
 from abc import abstractmethod
-from .ParameterSubspace import ParameterSubspace
+from xpsi.ParameterSubspace import ParameterSubspace
 
 class BackgroundError(xpsiError):
     """ Raised if there is an issue with the incident background . """
