+++ xpsi/Elsewhere.py
@@ -1,14 +1,13 @@
 from __future__ import division, print_function
 
-from .global_imports import *
-#from . import global_imports
+from xpsi.global_imports import *
 
-from .cellmesh.global_mesh import construct_closed_cellMesh as _construct_closed_cellMesh
-from .cellmesh.rays import compute_rays as _compute_rays
-from .cellmesh.integrator_for_time_invariance import integrate as _integrator
+from xpsi.cellmesh.global_mesh import construct_closed_cellMesh as _construct_closed_cellMesh
+from xpsi.cellmesh.rays import compute_rays as _compute_rays
+from xpsi.cellmesh.integrator_for_time_invariance import integrate as _integrator
 
-from .Parameter import Parameter
-from .ParameterSubspace import ParameterSubspace
+from xpsi.Parameter import Parameter
+from xpsi.ParameterSubspace import ParameterSubspace
 
 class RayError(xpsiError):
     """ Raised if a problem was encountered during ray integration. """
