+++ xpsi/HotRegion.py
@@ -1,12 +1,12 @@
-from .global_imports import *
+from xpsi.global_imports import *
 
-from .cellmesh.mesh_tools import allocate_cells as _allocate_cells
-from .cellmesh.mesh import construct_spot_cellMesh as _construct_spot_cellMesh
-from .cellmesh.polar_mesh import construct_polar_cellMesh as _construct_polar_cellMesh
-from .cellmesh.rays import compute_rays as _compute_rays
+from xpsi.cellmesh.mesh_tools import allocate_cells as _allocate_cells
+from xpsi.cellmesh.mesh import construct_spot_cellMesh as _construct_spot_cellMesh
+from xpsi.cellmesh.polar_mesh import construct_polar_cellMesh as _construct_polar_cellMesh
+from xpsi.cellmesh.rays import compute_rays as _compute_rays
 
-from .Parameter import Parameter, Derive
-from .ParameterSubspace import ParameterSubspace
+from xpsi.Parameter import Parameter, Derive
+from xpsi.ParameterSubspace import ParameterSubspace
 
 class RayError(xpsiError):
     """ Raised if a problem was encountered during ray integration. """
@@ -481,9 +481,9 @@ class HotRegion(ParameterSubspace):
 
         # find the required integrator
         if declaration: # can we safely assume azimuthal invariance?
-            from .cellmesh.integrator_for_azimuthal_invariance import integrate as _integrator
+            from xpsi.cellmesh.integrator_for_azimuthal_invariance import integrate as _integrator
         else: # more general purpose
-            from .cellmesh.integrator import integrate as _integrator
+            from xpsi.cellmesh.integrator import integrate as _integrator
         self._integrator = _integrator
 
     @property
