+++ xpsi/Instrument.py
@@ -1,10 +1,10 @@
 __all__ = ["Instrument"]
 
-from .global_imports import *
+from xpsi.global_imports import *
 
-from . import make_verbose
+from xpsi.__init__ import make_verbose
 
-from .ParameterSubspace import ParameterSubspace
+from xpsi.ParameterSubspace import ParameterSubspace
 
 class ResponseError(xpsiError):
     """ Raised if there is a problem with the input response matrix. """
