+++ xpsi/Interstellar.py
@@ -1,9 +1,9 @@
 __all__ = ["Interstellar"]
 
-from .global_imports import *
+from xpsi.global_imports import *
 
 from abc import abstractmethod
-from .ParameterSubspace import ParameterSubspace
+from xpsi.ParameterSubspace import ParameterSubspace
 
 class Interstellar(ParameterSubspace):
     """ Base class for model interstellar X-ray processes (e.g., attenuation)."""
