+++ xpsi/Parameter.py
@@ -1,6 +1,6 @@
-from .global_imports import *
+from xpsi.global_imports import *
 
-from . import make_verbose
+from xpsi import make_verbose
 
 import string
